move client version generation to new versionUtils.js
