+++ buildSrc/bump-version.js
@@ -5,6 +5,7 @@ import { Option, program } from "commander"
 import fs from "node:fs"
 import path from "node:path"
 import { $ } from "zx"
+import { calculateClientVersions } from "./versionUtils.js"
 
 await program
 	.addOption(new Option("-p, --platform <platform>", "version for which platform to bump").choices(["all", "webdesktop", "android", "ios"]).default("all"))
@@ -23,10 +24,11 @@ await program
  */
 async function run({ platform }) {
 	console.log(`bumping version for ${platform ?? "all"}`)
-	const currentVersionString = await readCurrentVersion()
-	const currentVersion = parseCurrentVersion(currentVersionString)
-	const newVersion = makeNewVersion(currentVersion)
-	const newVersionString = newVersion.join(".")
+	const clientVersions = await calculateClientVersions()
+
+	const currentVersion = clientVersions.currentVersion
+	const currentVersionString = clientVersions.currentVersionString
+	const newVersionString = clientVersions.newVersionString
 
 	if (platform === "all" || platform === "webdesktop") {
 		await bumpWorkspaces(newVersionString)
@@ -50,34 +52,6 @@ async function run({ platform }) {
 	}
 
 	console.log(`Bumped version ${currentVersionString} -> ${newVersionString}`)
-}
-
-async function readCurrentVersion() {
-	return JSON.parse(await fs.promises.readFile("./package.json", { encoding: "utf8" })).version
-}
-
-/**
- * Read versions of all models from generated ModelInfos.
- * @returns {number[]}
- */
-function readModelVersions() {
-	const appNames = fs.readdirSync("./src/api/entities/")
-	return appNames.map((appName) => {
-		const modelInfoString = fs.readFileSync(`./src/api/entities/${appName}/ModelInfo.ts`, { encoding: "utf8" })
-		const versionPrefix = "version:"
-		const versionNumberStart = modelInfoString.indexOf(versionPrefix) + versionPrefix.length
-		const version = modelInfoString.substring(versionNumberStart, modelInfoString.indexOf(",", versionNumberStart))
-		console.log(` > ${appName} version ${version}`)
-		return parseInt(version, 10)
-	})
-}
-
-/**
- * @param currentVersionString {string}
- * @return {number[]}
- */
-function parseCurrentVersion(currentVersionString) {
-	return currentVersionString.split(".").map((n) => parseInt(n, 10))
 }
 
 /**
@@ -128,33 +102,6 @@ async function bumpAndroidVersion(currentVersion, newVersionString) {
 		.replace(new RegExp(/versionCode (\d+)/), `versionCode ${newVersionCodeString}`)
 	console.log(`Bumped Android versionCode: ${oldVersionCodeString} -> ${newVersionCodeString}`)
 	await fs.promises.writeFile(buildGradlePath, newBuildGradleString)
-}
-
-/**
- * Creates a new client version number.
- * * Major is the sum of all current model versions.
- * * Minor is the current date
- * * Patch is always increased by one in case Major or Minor have not been changed
- * @param currentVersion {number[]}
- * @return {number[]}
- */
-function makeNewVersion(currentVersion) {
-	const modelVersions = readModelVersions()
-	const majorVersion = modelVersions.reduce((sum, current) => sum + current, 0)
-
-	const now = new Date()
-	const year = now.getFullYear().toString().substring(2, 4)
-	const month = now.getMonth() + 1
-	const formattedMonth = month >= 10 ? month : `0${month}`
-	const formattedDay = now.getDate() >= 10 ? now.getDate() : `0${now.getDate()}`
-	const minorVersion = parseInt(`${year}${formattedMonth}${formattedDay}`, 10)
-
-	const current = currentVersion.slice()
-	let patchVersion = 0
-	if (current[0] === majorVersion && current[1] === minorVersion) {
-		patchVersion = current[2] + 1
-	}
-	return [majorVersion, minorVersion, patchVersion]
 }
 
 /**
